requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==39.0.0
google-auth==2.48.0
gunicorn==25.1.0
idna==3.11
mysqlclient==2.2.8
packaging==26.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.4
tzdata==2025.3
urllib3==2.6.3
